Made the id for the CustomField model required
Updated CustomField to make the id required.
Updated CustomField to add another parameter to the constructor to
indicate if this model is new or loaded from the server, so we know if
we need to call POST or PUT since the model will always have an id.
Added unit tests to check the correct HTTP verb is used.
Refactored the base model and collection so that even though the
collection can rely on the endpoint method from the model, it does not
need to instantiate it.