Liquid / Legacy format structure
Rather then class stuff, split tables and liquid tag to read the cookie.

If there is no cookie or issue with the cookie OR before new interface implemented then both tables neatly separated will be shown. With liquid and the cookie working then the legacy and liquid will be shown depending on the cookie.